+++ src/main/java/de/samply/share/client/util/connector/CtsConnector.java
@@ -7,6 +7,7 @@ import com.google.gson.JsonObject;
 import com.jayway.jsonpath.Configuration;
 import com.jayway.jsonpath.JsonPath;
 import com.jayway.jsonpath.Option;
+import com.jayway.jsonpath.PathNotFoundException;
 import com.mchange.rmi.NotAuthorizedException;
 import com.sun.jersey.api.NotFoundException;
 import de.samply.common.http.HttpConnector;
@@ -30,6 +31,7 @@ import java.util.regex.Pattern;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 import org.apache.http.Consts;
+import org.apache.http.Header;
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpHeaders;
 import org.apache.http.HttpHost;
@@ -57,6 +59,8 @@ public class CtsConnector {
   private static final Logger logger = LogManager.getLogger(CtsConnector.class);
   private static final String CONTENT_TYPE_CTS_FHIR_JSON = "application/fhir+json; fhirVersion=4.0";
   private static final FhirResource fhirResource = new FhirResource();
+  private static final String X_BK_PSEUDONYM_JSONPATHS = "X-BK-pseudonym-jsonpaths";
+  private static final String X_BK_TARGET_URL = "X-BK-target-url";
   private transient HttpConnector httpConnector;
   private CloseableHttpClient httpClient;
   private String ctsBaseUrl;
@@ -156,7 +160,7 @@ public class CtsConnector {
     // Set up the API call
     HttpEntity entity = new StringEntity(pseudonimisedPatient.toString(), Consts.UTF_8);
     HttpPost httpPost = new HttpPost(ctsBaseUrl);
-    httpPost.setHeader(HttpHeaders.CONTENT_TYPE, CONTENT_TYPE_CTS_FHIR_JSON);
+    httpPost.setHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON);
     httpPost.setEntity(entity);
     CloseableHttpResponse response = null;
     try {
@@ -193,11 +197,22 @@ public class CtsConnector {
       HashMap<String, Object> headerMapToSend)
       throws IOException, IllegalArgumentException,
       NotFoundException, NotAuthorizedException {
-    String encryptedIds = readIds(patient,
-        httpHeaders.getRequestHeader("X-BK-pseudonym-jsonpaths").get(0), false);
+    List<String> urlTargetHeaders = httpHeaders.getRequestHeader(X_BK_TARGET_URL);
+    String encryptedIds = patient;//or a empty string
+    String urlTarget;
+    if (urlTargetHeaders != null && !urlTargetHeaders.isEmpty()) {
+      urlTarget = urlTargetHeaders.get(0);
+    } else {
+      logger.error("PostLocalPatientToCentralCts: X-BK-target-url is empty");
+      return Response.status(400).entity("X-BK-target-url is empty").build();
+    }
+    List<String> jsonpathsHeaders = httpHeaders.getRequestHeader(X_BK_PSEUDONYM_JSONPATHS);
+    if (jsonpathsHeaders != null && !jsonpathsHeaders.isEmpty()) {
+      encryptedIds = readIds(patient, jsonpathsHeaders.get(0), false);
+    }
     // Set up the API call
     HttpEntity entity = new StringEntity(encryptedIds, Consts.UTF_8);
-    HttpPost httpPost = new HttpPost(httpHeaders.getRequestHeader("X-BK-target-url").get(0));
+    HttpPost httpPost = new HttpPost(urlTarget);
     httpPost.setHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON);
     for (Entry<String, Object> entry : headerMapToSend.entrySet()) {
       httpPost.setHeader(entry.getKey(), entry.getValue().toString());
@@ -208,9 +223,12 @@ public class CtsConnector {
       response = httpClient.execute(httpPost);
       int statusCode = response.getStatusLine().getStatusCode();
       if (statusCode == 200 || statusCode == 201) {
-        String patients = readIds(patient,
-            httpHeaders.getRequestHeader("X-BK-pseudonym-jsonpaths").get(0), true);
-        return Response.status(statusCode).entity(patients).build();
+        String responseAsString = EntityUtils.toString(response.getEntity());
+        Header responseJsonpaths = response.getFirstHeader(X_BK_PSEUDONYM_JSONPATHS);
+        if (responseJsonpaths != null) {
+          responseAsString = readIds(responseAsString, responseJsonpaths.getValue(), true);
+        }
+        return getResponse(response, statusCode, responseAsString);
       }
       String message =
           "CTS server response: statusCode:" + statusCode + "; response: " + response.toString();
@@ -218,12 +236,36 @@ public class CtsConnector {
       if (responseBody != null && !responseBody.isEmpty()) {
         message += ";body: " + responseBody;
       }
+      logger.error("PostLocalPatientToCentralCts response: " + message);
       return Response.status(statusCode).entity(message).build();
     } catch (IOException e) {
       throw new IOException(e);
+    } catch (StringIndexOutOfBoundsException e) {
+      throw new StringIndexOutOfBoundsException(e.getMessage());
+    } catch (PathNotFoundException e) {
+      throw new PathNotFoundException(e.getMessage());
     } finally {
       closeResponse(response);
     }
+  }
+
+  /**
+   * build a response from the a original response, headers and status code.
+   *
+   * @param response CloseableHttpResponse
+   * @param statusCode status code from the response
+   * @param responseAsString response body
+   * @return http response
+   */
+  private Response getResponse(CloseableHttpResponse response, int statusCode,
+      String responseAsString) {
+    Response.ResponseBuilder builder = Response.status(statusCode).entity(responseAsString);
+    //headers
+    Header[] responseHeaders = response.getAllHeaders();
+    for (Header header : headersToFilter(responseHeaders)) {
+      builder.header(header.getName(), header.getValue());
+    }
+    return builder.build();
   }
 
   /**
@@ -238,7 +280,6 @@ public class CtsConnector {
     // Recycle the authorization cookies that we received from the CTS
     cookieStore.addCookie(ctsAuthorization.codeCookie);
     cookieStore.addCookie(ctsAuthorization.userCookie);
-
     HttpContext ctsContext = new BasicHttpContext();
     ctsContext.setAttribute(HttpClientContext.COOKIE_STORE, cookieStore);
 
@@ -337,14 +378,28 @@ public class CtsConnector {
   }
 
   private String readIds(String json, String headerIdKey, boolean response)
-      throws IOException, NotAuthorizedException {
+      throws IOException, NotAuthorizedException, StringIndexOutOfBoundsException,
+      PathNotFoundException {
     String headerIdKeyString = new String(Base64.getDecoder().decode(headerIdKey));
-    headerIdKeyString = headerIdKeyString.substring(headerIdKeyString.indexOf("$"),
-        headerIdKeyString.indexOf("\"]"));
-    String patientJson = json;
-    Configuration conf = Configuration.defaultConfiguration().addOptions(Option.ALWAYS_RETURN_LIST);
-    List<String> ids = JsonPath.using(conf).parse(patientJson).read(headerIdKeyString);
-    patientJson = replaceIdsWithEncryptedIds(patientJson, ids, response);
+    String patientJson = null;
+    try {
+      headerIdKeyString = headerIdKeyString.substring(headerIdKeyString.indexOf("$"),
+          headerIdKeyString.indexOf("\"]"));
+      patientJson = json;
+      Configuration conf = Configuration.defaultConfiguration()
+          .addOptions(Option.ALWAYS_RETURN_LIST);
+      List<String> ids = JsonPath.using(conf).parse(patientJson).read(headerIdKeyString);
+      patientJson = replaceIdsWithEncryptedIds(patientJson, ids, response);
+
+    } catch (StringIndexOutOfBoundsException e) {
+      logger.error("jsonpath does not match the expected format ($ and [] are expected)", e);
+      throw new StringIndexOutOfBoundsException(
+          "Jsonpath does not match the expected format ($ and [] are expected):  " + e);
+    } catch (PathNotFoundException e) {
+      logger.error("could not found isonpath matches", e);
+      throw new PathNotFoundException("could not find jsonpath matches:  " + e);
+    }
+
     return patientJson;
   }
 
@@ -430,6 +485,25 @@ public class CtsConnector {
       String bodyResponse) {
     return message + "; statusCode: " + statusCode + "; reason: " + reasonPhrase + ";body: "
         + bodyResponse;
+  }
+
+  /**
+   * Filter Headers to forward to local CTS.
+   *
+   * @param headers headers from response
+   * @return all headers with defined suffixes
+   */
+  private Header[] headersToFilter(Header[] headers) {
+    String[] headersToPropagate = {"x-cds-", "x-bk-"};
+    List<Header> headersToSend = new ArrayList<>();
+    for (Header header : headers) {
+      for (String headerToPropagate : headersToPropagate) {
+        if (header.getName().toLowerCase().startsWith(headerToPropagate)) {
+          headersToSend.add(header);
+        }
+      }
+    }
+    return headersToSend.toArray(new Header[headersToSend.size()]);
   }
 
   /**